Initial README
Adds the framework and many of the endpoint instructions. Delete palettes is not working without the id and the id is not a required parameter so we need to grab fix that. I also added the id to the message when a project is deleted. This will change the delete response for projects, once that change is merged.